refactoring prompt
